refactor: response 폴더 이동
